+++ src/collectives/all_reduce.cu
@@ -66,7 +66,7 @@ __global__ void AllReduceKernel(const KernelArgs<T> args) {
   }
   __syncthreads();
 
-  int step = 0;
+  uint64_t step = 0ULL;
   int poffset, noffset = 0;
 
   // Compute pointers
@@ -197,8 +197,8 @@ __global__ void AllReduceKernel(const KernelArgs<T> args) {
   if (tid == 0) {
     // Wait for next to have consumed all data before we reset the flag
     waitDoneFromNext.wait(NUM_SUBSTEPS*(step + NUM_BUFCHUNKS));
-    *ring->send.conn.head = 0;
-    *ring->recv.conn.tail = 0;
+    *ring->send.conn.head = 0ULL;
+    *ring->recv.conn.tail = 0ULL;
     __threadfence_system();
     *ring->recv.conn.opCount = args.opCount+1;
   }
