+++ src/collectives/broadcast.cu
@@ -66,7 +66,7 @@ __global__ void BroadcastKernel(const KernelArgs<T> args) {
   }
   __syncthreads();
   
-  int step = 0;
+  uint64_t step = 0ULL;
   int boffset = 0;
 
   // Compute pointers
@@ -136,9 +136,9 @@ __global__ void BroadcastKernel(const KernelArgs<T> args) {
     if (nextRank != root) { 
       // Wait for next to have consumed data before resetting the flag
       waitDoneFromNext.wait(NUM_SUBSTEPS*(step + NUM_BUFCHUNKS - 1));
-      *ring->send.conn.head = 0;
+      *ring->send.conn.head = 0ULL;
     }
-    *ring->recv.conn.tail = 0;
+    *ring->recv.conn.tail = 0ULL;
     __threadfence_system();
     *ring->recv.conn.opCount = args.opCount+1;
   }
